requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
jupyter>=1.0.0
notebook>=7.0.0
nbformat>=5.9.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pydantic>=2.0.0
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
mypy>=1.5.0
colorama>=0.4.6  # For colored terminal output 